dirname(__file__) => __dir__ \r\n => \n Мелкие ошибки и описки

+++ addons/admin/modules/misc_ajax.php
@@ -12,22 +12,28 @@ if(!defined('CMS'))die;
 
 $event=isset($_POST['event']) ? (string)$_POST['event'] : '';
 switch($event)
-{	case'fixed':
+{
+	case'fixed':
 		$log=isset($_POST['log']) ? (string)$_POST['log'] : '';
 		if(in_array($log,array('errors','db_errors','request_errors')))
-		{			$id=isset($_POST['id']) ? (string)$_POST['id'] : '';			$log=Eleanor::$root.'addons/logs/'.$log.'.log';
+		{
+			$id=isset($_POST['id']) ? (string)$_POST['id'] : '';
+			$log=Eleanor::$root.'addons/logs/'.$log.'.log';
 			$hlog=$log.'.inc';
 
 			if(is_file($log) and is_file($hlog))
-			{				$info=(array)unserialize(file_get_contents($hlog));
+			{
+				$info=(array)unserialize(file_get_contents($hlog));
 				if(isset($info[$id]))
-				{					if(count($info)==1)
+				{
+					if(count($info)==1)
 					{
 						Files::Delete($log);
 						Files::Delete($hlog);
 					}
 					else
-					{						$fh=fopen($log,'rb+');
+					{
+						$fh=fopen($log,'rb+');
 						if(flock($fh,LOCK_EX))
 						{
 							$diff=Files::FReplace($fh,'',$info[$id]['o'],$info[$id]['l']+strlen(PHP_EOL)*2);
@@ -40,13 +46,17 @@ switch($event)
 							file_put_contents($hlog,serialize($info));
 						}
 						else
-						{							fclose($fh);
+						{
+							fclose($fh);
 							Error();
-							break;						}
+							break;
+						}
 					}
-				}
+
+				}
 			}
-			Result(true);
+
+			Result(true);
 			break;
 		}
 	default:
